backend/requirements.txt
# 后端服务依赖
flask>=2.0.0
flask-cors>=3.0.0

# 爬虫依赖
requests>=2.25.0
beautifulsoup4>=4.9.0

# 配置文件依赖
pyyaml>=6.0.0

# 定时任务依赖
schedule>=1.1.0
